raspberry-pi/requirements.txt
oauthlib==1.0.0
python-dateutil==2.4.2
python-firebase==1.2
pyusb==1.0.0b1
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
twython==3.3.0
wheel==0.24.0
